requirement.txt
tensorflow==2.19.0
numpy==2.1.3
opencv-contrib-python==4.11.0.86
pillow==11.2.1
pandas==2.3.0
matplotlib==3.10.3
scikit-learn==1.7.0
ipython==8.37.0
bing-image-downloader==1.1.2
tk
keras==3.10.0
keras-vggface==0.6
mediapipe==0.10.21
seaborn==0.13.2